ChatBot/requirements.txt
altair<5
streamlit
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
langchain_openai
python-dotenv